remove calc_gcd() from brain_prime
